Analysis of Music Listening Habits & Mental Health Outcomes
CSV File of the results of a survey conducted. The survey asked participants about various music consumption and creation habits. It also asked them to rate their own mental health on a scale of 0-10 with regards to anxiety, insomnia, depression and OCD; a higher score meant higher reported levels of that particular mental health outcome.

Python Script file is an analysis of the data in the CSV file and a report of any notable trends or findings.